+++ schema/schema_test.go
@@ -44,37 +44,39 @@ resource "aws_instance2" "web" {
 					File: "test.tf",
 					Resources: []*Resource{
 						{
-							File: "test.tf",
 							Type: "aws_instance2",
 							Id:   "web",
-							Pos: token.Pos{
-								Filename: "test.tf",
-								Offset:   258,
-								Line:     15,
-								Column:   32,
-							},
-							Attrs: map[string]*Attribute{
-								"root_block_device": {
-									Poses: []token.Pos{
-										{
-											Filename: "test.tf",
-											Offset:   282,
-											Line:     16,
-											Column:   23,
+							Source: &Source{
+								File: "test.tf",
+								Pos: token.Pos{
+									Filename: "test.tf",
+									Offset:   258,
+									Line:     15,
+									Column:   32,
+								},
+								Attrs: map[string]*Attribute{
+									"root_block_device": {
+										Poses: []token.Pos{
+											{
+												Filename: "test.tf",
+												Offset:   282,
+												Line:     16,
+												Column:   23,
+											},
 										},
-									},
-									Vals: []interface{}{
-										map[string]interface{}{
-											"volume_size": token.Token{
-												Type: 9,
-												Pos: token.Pos{
-													Filename: "test.tf",
-													Offset:   302,
-													Line:     17,
-													Column:   19,
+										Vals: []interface{}{
+											map[string]interface{}{
+												"volume_size": token.Token{
+													Type: 9,
+													Pos: token.Pos{
+														Filename: "test.tf",
+														Offset:   302,
+														Line:     17,
+														Column:   19,
+													},
+													Text: "\"24\"",
+													JSON: false,
 												},
-												Text: "\"24\"",
-												JSON: false,
 											},
 										},
 									},
@@ -82,83 +84,85 @@ resource "aws_instance2" "web" {
 							},
 						},
 						{
-							File: "test.tf",
 							Type: "aws_instance",
 							Id:   "web",
-							Pos: token.Pos{
-								Filename: "test.tf",
-								Offset:   31,
-								Line:     2,
-								Column:   31,
-							},
-							Attrs: map[string]*Attribute{
-								"ami": {
-									Poses: []token.Pos{
-										{
-											Filename: "test.tf",
-											Offset:   51,
-											Line:     3,
-											Column:   19,
-										},
-									},
-									Vals: []interface{}{
-										token.Token{
-											Type: 9,
-											Pos: token.Pos{
+							Source: &Source{
+								File: "test.tf",
+								Pos: token.Pos{
+									Filename: "test.tf",
+									Offset:   31,
+									Line:     2,
+									Column:   31,
+								},
+								Attrs: map[string]*Attribute{
+									"ami": {
+										Poses: []token.Pos{
+											{
 												Filename: "test.tf",
 												Offset:   51,
 												Line:     3,
 												Column:   19,
 											},
-											Text: "\"ami-b73b63a0\"",
-											JSON: false,
+										},
+										Vals: []interface{}{
+											token.Token{
+												Type: 9,
+												Pos: token.Pos{
+													Filename: "test.tf",
+													Offset:   51,
+													Line:     3,
+													Column:   19,
+												},
+												Text: "\"ami-b73b63a0\"",
+												JSON: false,
+											},
 										},
 									},
-								},
-								"instance_type": {
-									Poses: []token.Pos{
-										{
-											Filename: "test.tf",
-											Offset:   84,
-											Line:     4,
-											Column:   19,
-										},
-									},
-									Vals: []interface{}{
-										token.Token{
-											Type: 9,
-											Pos: token.Pos{
+									"instance_type": {
+										Poses: []token.Pos{
+											{
 												Filename: "test.tf",
 												Offset:   84,
 												Line:     4,
 												Column:   19,
 											},
-											Text: "\"t1.2xlarge\"",
-											JSON: false,
 										},
-									},
-								},
-								"tags": {
-									Poses: []token.Pos{
-										{
-											Filename: "test.tf",
-											Offset:   121,
-											Line:     6,
-											Column:   8,
-										},
-									},
-									Vals: []interface{}{
-										map[string]interface{}{
-											"Name": token.Token{
+										Vals: []interface{}{
+											token.Token{
 												Type: 9,
 												Pos: token.Pos{
 													Filename: "test.tf",
-													Offset:   134,
-													Line:     7,
-													Column:   12,
+													Offset:   84,
+													Line:     4,
+													Column:   19,
 												},
-												Text: "\"HelloWorld\"",
+												Text: "\"t1.2xlarge\"",
 												JSON: false,
+											},
+										},
+									},
+									"tags": {
+										Poses: []token.Pos{
+											{
+												Filename: "test.tf",
+												Offset:   121,
+												Line:     6,
+												Column:   8,
+											},
+										},
+										Vals: []interface{}{
+											map[string]interface{}{
+												"Name": token.Token{
+													Type: 9,
+													Pos: token.Pos{
+														Filename: "test.tf",
+														Offset:   134,
+														Line:     7,
+														Column:   12,
+													},
+													Text: "\"HelloWorld\"",
+													JSON: false,
+												},
 											},
 										},
 									},
@@ -166,48 +170,50 @@ resource "aws_instance2" "web" {
 							},
 						},
 						{
-							File: "test.tf",
 							Type: "aws_instance",
 							Id:   "web2",
-							Pos: token.Pos{
-								Filename: "test.tf",
-								Offset:   185,
-								Line:     11,
-								Column:   32,
-							},
-							Attrs: map[string]*Attribute{
-								"security_groups": {
-									Poses: []token.Pos{
-										{
-											Filename: "test.tf",
-											Offset:   207,
-											Line:     12,
-											Column:   21,
-										},
-									},
-									Vals: []interface{}{
-										[]interface{}{
-											token.Token{
-												Type: 9,
-												Pos: token.Pos{
-													Filename: "test.tf",
-													Offset:   208,
-													Line:     12,
-													Column:   22,
-												},
-												Text: "\"sg-1\"",
-												JSON: false,
+							Source: &Source{
+								File: "test.tf",
+								Pos: token.Pos{
+									Filename: "test.tf",
+									Offset:   185,
+									Line:     11,
+									Column:   32,
+								},
+								Attrs: map[string]*Attribute{
+									"security_groups": {
+										Poses: []token.Pos{
+											{
+												Filename: "test.tf",
+												Offset:   207,
+												Line:     12,
+												Column:   21,
 											},
-											token.Token{
-												Type: 9,
-												Pos: token.Pos{
-													Filename: "test.tf",
-													Offset:   216,
-													Line:     12,
-													Column:   30,
+										},
+										Vals: []interface{}{
+											[]interface{}{
+												token.Token{
+													Type: 9,
+													Pos: token.Pos{
+														Filename: "test.tf",
+														Offset:   208,
+														Line:     12,
+														Column:   22,
+													},
+													Text: "\"sg-1\"",
+													JSON: false,
 												},
-												Text: "\"sg-2\"",
-												JSON: false,
+												token.Token{
+													Type: 9,
+													Pos: token.Pos{
+														Filename: "test.tf",
+														Offset:   216,
+														Line:     12,
+														Column:   30,
+													},
+													Text: "\"sg-2\"",
+													JSON: false,
+												},
 											},
 										},
 									},
@@ -252,83 +258,85 @@ resource "aws_instance" "web" {
 					File: "test.tf",
 					Resources: []*Resource{
 						{
-							File: "test.tf",
 							Type: "aws_instance",
 							Id:   "web",
-							Pos: token.Pos{
-								Filename: "test.tf",
-								Offset:   31,
-								Line:     2,
-								Column:   31,
-							},
-							Attrs: map[string]*Attribute{
-								"ami": {
-									Poses: []token.Pos{
-										{
-											Filename: "test_override.tf",
-											Offset:   51,
-											Line:     3,
-											Column:   19,
-										},
-									},
-									Vals: []interface{}{
-										token.Token{
-											Type: 9,
-											Pos: token.Pos{
+							Source: &Source{
+								File: "test.tf",
+								Pos: token.Pos{
+									Filename: "test.tf",
+									Offset:   31,
+									Line:     2,
+									Column:   31,
+								},
+								Attrs: map[string]*Attribute{
+									"ami": {
+										Poses: []token.Pos{
+											{
 												Filename: "test_override.tf",
 												Offset:   51,
 												Line:     3,
 												Column:   19,
 											},
-											Text: "\"ami-override\"",
-											JSON: false,
+										},
+										Vals: []interface{}{
+											token.Token{
+												Type: 9,
+												Pos: token.Pos{
+													Filename: "test_override.tf",
+													Offset:   51,
+													Line:     3,
+													Column:   19,
+												},
+												Text: "\"ami-override\"",
+												JSON: false,
+											},
 										},
 									},
-								},
-								"instance_type": {
-									Poses: []token.Pos{
-										{
-											Filename: "test_override.tf",
-											Offset:   84,
-											Line:     4,
-											Column:   19,
-										},
-									},
-									Vals: []interface{}{
-										token.Token{
-											Type: 9,
-											Pos: token.Pos{
+									"instance_type": {
+										Poses: []token.Pos{
+											{
 												Filename: "test_override.tf",
 												Offset:   84,
 												Line:     4,
 												Column:   19,
 											},
-											Text: "\"t2.nano\"",
-											JSON: false,
 										},
-									},
-								},
-								"tags": {
-									Poses: []token.Pos{
-										{
-											Filename: "test_override.tf",
-											Offset:   102,
-											Line:     6,
-											Column:   8,
-										},
-									},
-									Vals: []interface{}{
-										map[string]interface{}{
-											"Version": token.Token{
+										Vals: []interface{}{
+											token.Token{
 												Type: 9,
 												Pos: token.Pos{
 													Filename: "test_override.tf",
-													Offset:   118,
-													Line:     7,
-													Column:   15,
+													Offset:   84,
+													Line:     4,
+													Column:   19,
 												},
-												Text: "\"0.1\"",
+												Text: "\"t2.nano\"",
 												JSON: false,
+											},
+										},
+									},
+									"tags": {
+										Poses: []token.Pos{
+											{
+												Filename: "test_override.tf",
+												Offset:   102,
+												Line:     6,
+												Column:   8,
+											},
+										},
+										Vals: []interface{}{
+											map[string]interface{}{
+												"Version": token.Token{
+													Type: 9,
+													Pos: token.Pos{
+														Filename: "test_override.tf",
+														Offset:   118,
+														Line:     7,
+														Column:   15,
+													},
+													Text: "\"0.1\"",
+													JSON: false,
+												},
 											},
 										},
 									},
